+++ app/api/Datasets.scala
@@ -204,7 +204,7 @@ class  Datasets @Inject()(
       appConfig.incrementCount('datasets, 1)
 
       //event will be added whether creation is success.
-      events.addObjectEvent(request.user, d.id, d.name, "create_dataset")
+      events.addObjectEvent(request.user, d.id, d.name, EventType.CREATE_DATASET.toString)
       datasets.index(d.id)
 
       (request.body \ "file_id").asOpt[String] match {
@@ -215,7 +215,7 @@ class  Datasets @Inject()(
                 case Some(id) => {
                   d.spaces.map( spaceId => spaces.get(spaceId)).flatten.map{ s =>
                     spaces.addDataset(d.id, s.id)
-                    events.addSourceEvent(request.user, d.id, d.name, s.id, s.name, "add_dataset_space")
+                    events.addSourceEvent(request.user, d.id, d.name, s.id, s.name, EventType.ADD_DATASET_SPACE.toString)
                   }
                   attachExistingFileHelper(UUID(id), file.id, d, file, request.user)
                   files.index(UUID(file_id))
@@ -294,7 +294,7 @@ class  Datasets @Inject()(
           }
           case None => InternalServerError("User Not found")
         }
-        events.addObjectEvent(request.user, d.id, d.name, "create_dataset")
+        events.addObjectEvent(request.user, d.id, d.name, EventType.CREATE_DATASET.toString)
 
         datasets.insert(d) match {
           case Some(id) => {
@@ -306,7 +306,7 @@ class  Datasets @Inject()(
             Logger.debug("About to call addDataset on spaces service")
             d.spaces.map( spaceId => spaces.get(spaceId)).flatten.map{ s =>
               spaces.addDataset(d.id, s.id)
-              events.addSourceEvent(request.user, d.id, d.name, s.id, s.name, "add_dataset_space")
+              events.addSourceEvent(request.user, d.id, d.name, s.id, s.name, EventType.ADD_DATASET_SPACE.toString)
 
             }
             //Add this dataset to a collection if needed
@@ -615,6 +615,8 @@ class  Datasets @Inject()(
           p.metadataAddedToResource(metadataId, metadata.attachedTo, mdMap, Utils.baseUrl(request))
         }
 
+        events.addObjectEvent(None, id, x.name, EventType.ADD_METADATA_DATASET.toString)
+
         datasets.index(id)
         Ok(toJson(Map("status" -> "success")))
       }
@@ -672,6 +674,8 @@ class  Datasets @Inject()(
                     current.plugin[RabbitmqPlugin].foreach { p =>
                       p.metadataAddedToResource(metadataId, metadata.attachedTo, mdMap, Utils.baseUrl(request))
                     }
+
+                    events.addObjectEvent(None, id, x.name,EventType.ADD_METADATA_DATASET.toString)
 
                     datasets.index(id)
                     Ok(toJson("Metadata successfully added to db"))
@@ -776,7 +780,7 @@ class  Datasets @Inject()(
 
     datasets.get(id) match {
       case Some(dataset) => {
-        events.addObjectEvent(user, id, dataset.name, "addMetadata_dataset")
+        events.addObjectEvent(user, id, dataset.name, EventType.ADD_METADATA_DATASET.toString)
       }
     }
 
@@ -1010,7 +1014,7 @@ class  Datasets @Inject()(
       datasets.updateInformation(id, description, name)
       datasets.get(id) match {
         case Some(dataset) => {
-          events.addObjectEvent(user, id, dataset.name, "update_dataset_information")
+          events.addObjectEvent(user, id, dataset.name, EventType.UPDATE_DATASET_INFORMATION.toString)
         }
       }
       datasets.index(id)
@@ -1046,7 +1050,7 @@ class  Datasets @Inject()(
       datasets.updateName(id, name)
       datasets.get(id) match {
         case Some(dataset) => {
-          events.addObjectEvent(user, id, dataset.name, "update_dataset_information")
+          events.addObjectEvent(user, id, dataset.name, EventType.UPDATE_DATASET_INFORMATION.toString)
           datasets.index(id)
           // file in this dataset need to be indexed as well since dataset name will show in file list
           dataset.files.map(files.index(_))
@@ -1085,7 +1089,7 @@ class  Datasets @Inject()(
       datasets.updateDescription(id, description)
       datasets.get(id) match {
         case Some(dataset) => {
-          events.addObjectEvent(user, id, dataset.name, "update_dataset_information")
+          events.addObjectEvent(user, id, dataset.name, EventType.UPDATE_DATASET_INFORMATION.toString)
         }
       }
       datasets.index(id)
@@ -1122,7 +1126,7 @@ class  Datasets @Inject()(
       datasets.addCreator(id, creator)
       datasets.get(id) match {
         case Some(dataset) => {
-          events.addObjectEvent(user, id, dataset.name, "update_dataset_information")
+          events.addObjectEvent(user, id, dataset.name, EventType.UPDATE_DATASET_INFORMATION.toString)
         }
       }
       datasets.index(id)
@@ -1144,7 +1148,7 @@ class  Datasets @Inject()(
       datasets.removeCreator(id, creator)
       datasets.get(id) match {
         case Some(dataset) => {
-          events.addObjectEvent(user, id, dataset.name, "update_dataset_information")
+          events.addObjectEvent(user, id, dataset.name, EventType.UPDATE_DATASET_INFORMATION.toString)
         }
       }
       datasets.index(id)
@@ -1165,7 +1169,7 @@ class  Datasets @Inject()(
       datasets.moveCreator(id, creator, newPos)
       datasets.get(id) match {
         case Some(dataset) => {
-          events.addObjectEvent(user, id, dataset.name, "update_dataset_information")
+          events.addObjectEvent(user, id, dataset.name, EventType.UPDATE_DATASET_INFORMATION.toString)
         }
       }
       datasets.index(id)
@@ -1404,7 +1408,7 @@ class  Datasets @Inject()(
           datasets.addTags(id, userOpt, extractorOpt, tagsCleaned)
           datasets.get(id) match {
             case Some(dataset) => {
-              events.addObjectEvent(request.user, id, dataset.name, "add_tags_dataset")
+              events.addObjectEvent(request.user, id, dataset.name, EventType.ADD_TAGS_DATASET.toString)
             }
           }
           datasets.index(id)
@@ -1441,7 +1445,7 @@ class  Datasets @Inject()(
           datasets.removeTags(id, userOpt, extractorOpt, tagsCleaned)
           datasets.get(id) match {
             case Some(dataset) => {
-              events.addObjectEvent(request.user, id, dataset.name, "remove_tags_dataset")
+              events.addObjectEvent(request.user, id, dataset.name, EventType.REMOVE_TAGS_DATASET.toString)
             }
           }
           datasets.index(id)
@@ -1582,7 +1586,7 @@ class  Datasets @Inject()(
             datasets.index(id)
             datasets.get(id) match {
               case Some(dataset) => {
-                events.addSourceEvent(request.user, comment.id, comment.text , dataset.id, dataset.name, "add_comment_dataset")
+                events.addSourceEvent(request.user, comment.id, comment.text , dataset.id, dataset.name, EventType.ADD_COMMENT_DATASET.toString)
               }
             }
             Ok(comment.id.toString())
@@ -1778,7 +1782,7 @@ class  Datasets @Inject()(
           case "yes" => rdfsparql.removeDatasetFromGraphs(id)
           case _ => Logger.debug("userdfSPARQLStore not enabled")
         }
-        events.addObjectEvent(request.user, dataset.id, dataset.name, "delete_dataset")
+        events.addObjectEvent(request.user, dataset.id, dataset.name, EventType.DELETE_DATASET.toString)
         datasets.removeDataset(id, Utils.baseUrl(request))
         appConfig.incrementCount('datasets, -1)
 
@@ -1836,7 +1840,7 @@ class  Datasets @Inject()(
       case Some(u) => {
         val trashDatasets = datasets.listUserTrash(request.user,0)
         for (ds <- trashDatasets){
-          events.addObjectEvent(request.user, ds.id, ds.name, "delete_dataset")
+          events.addObjectEvent(request.user, ds.id, ds.name, EventType.DELETE_DATASET.toString)
           datasets.removeDataset(ds.id, Utils.baseUrl(request))
           appConfig.incrementCount('datasets, -1)
           current.plugin[ElasticsearchPlugin].foreach {
@@ -2007,7 +2011,7 @@ class  Datasets @Inject()(
         case Some(loggedInUser) => {
           datasets.get(id) match {
             case Some(dataset) => {
-              events.addObjectEvent(user, id, dataset.name, "follow_dataset")
+              events.addObjectEvent(user, id, dataset.name, EventType.FOLLOW_DATASET.toString)
               datasets.addFollower(id, loggedInUser.id)
               userService.followDataset(loggedInUser.id, id)
 
@@ -2035,7 +2039,7 @@ class  Datasets @Inject()(
       case Some(loggedInUser) => {
         datasets.get(id) match {
           case Some(dataset) => {
-            events.addObjectEvent(user, id, dataset.name, "unfollow_dataset")
+            events.addObjectEvent(user, id, dataset.name, EventType.UNFOLLOW_DATASET.toString)
             datasets.removeFollower(id, loggedInUser.id)
             userService.unfollowDataset(loggedInUser.id, id)
             Ok
@@ -2485,7 +2489,7 @@ class  Datasets @Inject()(
         datasets.get(id) match {
           case Some(dataset) if !dataset.isTRIAL => {
             datasets.update(dataset.copy(status = access))
-            events.addObjectEvent(user, id, dataset.name, "update_dataset_information")
+            events.addObjectEvent(user, id, dataset.name, EventType.UPDATE_DATASET_INFORMATION.toString)
             Ok(toJson(Map("status" -> "success")))
           }
           // If the dataset wasn't found by ID
@@ -2530,7 +2534,7 @@ class  Datasets @Inject()(
                       copyDatasetMetadata(dataset.id,UUID(id))
                       spaces.addDataset(d.id, spaceId)
                       relations.add(Relation(source = Node(datasetId.stringify, ResourceType.dataset), target = Node(d.id.stringify, ResourceType.dataset)))
-                      events.addSourceEvent(request.user, d.id, d.name, space.id, space.name, "add_dataset_space")
+                      events.addSourceEvent(request.user, d.id, d.name, space.id, space.name, EventType.ADD_DATASET_SPACE.toString)
 
                       dataset.folders.map { folder =>
                         copyFolders(folder, datasetId, "dataset", d.id)